% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_date.R
\name{annotation_date}
\alias{annotation_date}
\title{Adds date annotations to ggepicurve plots}
\usage{
annotation_date(
  date_unit = "month",
  y_percentage = 0.075,
  extra_lines = 0,
  text_par = list(),
  line_par = list(),
  format = "\%Y-\%m-\%d",
  fixed_y_scale = TRUE
)
}
\arguments{
\item{date_unit}{character string setting the unit of time for which to add the annotations.}

\item{y_percentage}{Numeric value specifying by which percentage of the complete Y-axis the annotations should be shifted.
This value will most likely need fine tuning for each graphic device but is independent of the number of cases.}

\item{extra_lines}{Numeric value (i.e. unit lines) specifying how much extra space for the annotations should be added at the bottom of the plot.
If annotation_date is added multiple times (e.g. for months and years) the last extra_lines value overwrites all previous ones.}

\item{text_par}{List of additional parameters for the text part of the annotation.}

\item{line_par}{List of additional parameters for the line-separator part of the annotation.}

\item{format}{Character string passed to \code{\link[=format.Date]{format.Date()}} specifying how the date for the annotation should be formatted.}

\item{fixed_y_scale}{Only relevant when creating a faceted plot, e.g. with \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}. Needs to be FALSE if Y-scale is set to free.}
}
\description{
Adds extra lines at the bottom of the plot or the panel and prints date annotations for calendar-weeks, months, quarters or years.
}
\details{
This function is designed to work with plots generated by \code{\link[=ggepicurve]{ggepicurve()}}.
However any ggplot2 plot with x-scale breaks consisting of character format dates could potentially work.
}
