% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_donnees.R
\name{plot_ggcurve}
\alias{plot_ggcurve}
\title{Make a ggplot from outbreak data}
\usage{
plot_ggcurve(df)
}
\arguments{
\item{df}{A data.frame prepared with the prepare_df function}
}
\value{
A ggplot object
}
\description{
Make a ggplot from outbreak data
}
\examples{
df_outbreak <- prepare_df(dates = outbreak_sim$dates, group_color = outbreak_sim$yard)
plot_ggcurve(df_outbreak)
}
