% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_donnees.R
\name{prepare_df}
\alias{prepare_df}
\title{Create an object df_epidemic}
\usage{
prepare_df(dates, group_color, groupe_facet = NA, period_limit = "week",
  id = NULL)
}
\arguments{
\item{dates}{Vector of dates}

\item{group_color}{Vector of character}

\item{groupe_facet}{Vector of character}

\item{period_limit}{A character vector of size 1. Either "day", "week" (default) or "month".}

\item{id}{Vector of character. The id of each obs. If NULL, then id will be autoassigned in the order of the dates.}
}
\value{
A data.frame
}
\description{
Create an object df_epidemic
}
