% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated case based epidemiological data set}
\format{
A data frame with
\describe{
\item{date}{Date of infection}
\item{sex}{Sex of infected individual}
\item{death}{Death status of infected individual}
\item{year}{Year of infection}
\item{region}{Region of infection}
}
}
\usage{
sim_data
}
\description{
53288 cases of simulated infections between 2010 and 2013.
}
\keyword{datasets}
