% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_outbreak}
\alias{simulate_outbreak}
\title{Simulate outbreak data}
\usage{
simulate_outbreak(n_obs = 60, date_range = as.Date(x = c("2016-12-01",
  "2017-01-30")))
}
\value{
A data.frame
}
\description{
Simulate outbreak data
}
