% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_date.R
\name{annotation_date}
\alias{annotation_date}
\title{Adds date annotations to ggepicurve plots}
\usage{
annotation_date(
  date_unit = "month",
  y_percentage = 0.075,
  y_scale = NULL,
  extra_lines = 0,
  text_par = list(),
  line_par = list(),
  format = "\%Y-\%m-\%d"
)
}
\arguments{
\item{date_unit}{character string setting the unit of time for which to add the annotations.}

\item{y_percentage}{Numeric value specifying by which percentage of the complete y-axis range the annotations should be shifted.}

\item{extra_lines}{Numeric value specifying how many lines should be added at the bottom of the plot.}

\item{text_par}{List of additional paramaters for the text part of the annotation.}

\item{line_par}{List of additional paramaters for the line-separator part of the annotation.}

\item{format}{Character string passed to \code{\link[=format.Date]{format.Date()}} specifying how the date for the annotation should be formated.}
}
\description{
Adds extra lines at the bottom of the plot or the panel and prints date annotations for calendar-weeks, months, quarters or years.
This function is designed to work with plots generated by \code{\link[=ggepicurve]{ggepicurve()}}.
However any ggplot2 plot with x-scale breaks consisting of character format dates could potentially work.
}
