% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggepicurve.R
\name{ggepicurve}
\alias{ggepicurve}
\title{Create a ggepicurve plot}
\usage{
ggepicurve(
  df_cases,
  date = "1. pos. LaborDiagnose",
  epi_squares = FALSE,
  fill_aes = NULL,
  label_aes = NULL,
  time_unit = "day",
  time_breaks = function(x) x,
  time_labels = function(x) x,
  start_date = NULL,
  end_date = NULL,
  gg_first = NULL,
  color_scheme = NULL,
  bar_par = list()
)
}
\description{
Create a ggepicurve plot
}
