% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggepicurve.R
\name{ggepicurve}
\alias{ggepicurve}
\title{Create a ggepicurve plot}
\usage{
ggepicurve(
  data,
  date,
  mapping = list(),
  epi_blocks = FALSE,
  date_unit = "day",
  date_labels = function(x) x,
  start_date = NULL,
  end_date = NULL,
  gg_first = NULL,
  col_par = list(),
  x_scale = NULL,
  suggested_theme = TRUE
)
}
\arguments{
\item{data}{Data set in which each line represents a case. Has to contain column of class date. For aggregated data (e.g. weekly cases) any date of the aggregation period (e.g. monday) can be used.}

\item{date}{Name of column which contains date by which cases are to be plotted.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}} or \code{\link[ggplot2:aes_]{ggplot2::aes_()}} for \code{\link[ggplot2:geom_bar]{ggplot2::geom_col()}}. This will not count as the default mapping at the top level of the plot.}

\item{epi_blocks}{Schould each case be plotted as a separate block? Can be TRUE, FALSE or a numeric. If it's a numeric, blocks will be drawn if the highest case count of the panel doesn't surpass its value.}

\item{date_unit}{Should the data be aggregated? Defaults to "day" (i.e. no aggregation). Other possible values are "week", "month", "quarter" and "year".
All cases of the respective date unit will be aggregated on the X-axis to its first day (except for "week" in which case the Thursday of each week will be the aggregation point).}

\item{date_labels}{Will be passed to \link[ggplot2:scale_discrete]{ggplot2::scale_x_discrete} as the labels argument. If a function is passed, the input will be the dates on the X-axis as a \strong{character} vector.}

\item{start_date}{Start date of X-axis. Character string in "YYYY-MM-DD" format.}

\item{end_date}{End date of X-axis. Character string in "YYYY-MM-DD" format.}

\item{gg_first}{List of gg objects to plot before (backgorund of) the epicurve.}

\item{col_par}{Additional parameters for the geom_col call.}

\item{suggested_theme}{Should suggested \code{\link[ggplot2:theme]{ggplot2::theme()}} adjustments be performed? Defaults to TRUE. See Details.}

\item{facet_x_scale}{Should be only used when creating a faceted plot with a free x scale.}
}
\value{
ggepicurve returns a gg object consisiting of a call to at least \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}, \code{\link[ggplot2:scale_discrete]{ggplot2::scale_x_discrete()}}, \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}} and \code{\link[ggepicurve:::stat_epicurve]{ggepicurve::::stat_epicurve()}}.
Additional gg objects can be added as usual with "+" but are not guaranteed to work. (e.g. \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} should work properly, but \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} often causes problems)
}
\description{
ggepicurve takes a data set of case based epidemiological data. This data set has to contain one column of the class \link{Date} by which the curve (actually columns) will be plotted.
This column is internally transformed into a factor so keep that in mind for any transformations of the X-axis labels.
}
\details{
\strong{suggested_theme} adjusts the following theme elements:
text = element_text(size = 12), panel.border = element_blank(), panel.background = element_rect(fill = NA), panel.grid.major.x = element_blank(), panel.grid.minor.x = element_blank(),
panel.grid.major.y = element_line(colour = "lightgrey"), panel.grid.minor.y = element_blank(), plot.title = element_text(face = "bold", vjust = 0.5, hjust = 0.5), axis.title = element_text(face = "bold"), axis.text.x = element_text(angle = 40, hjust = 1, vjust = 1), legend.position = "top", legend.direction = "horizontal")
}
\seealso{
\code{\link[=annotation_date]{annotation_date()}} for creating pretty X-axis annotations based on week, month, quarter or year.
}
